LayoutTests/platform/gtk/http/tests/security/window-named-proto-expected.txt

CONSOLE MESSAGE: line 1: TypeError: 'null' is not an object (evaluating 'document.body.innerHTML')
